requirements.txt
pyqt
requests
responses
pandas
numpy
